#10 Fix routes to v6.4
