+++ src/App.js
@@ -1,4 +1,4 @@
-import { Routes, Route } from 'react-router-dom';
+import { Outlet } from 'react-router-dom';
 
 import { inject } from '@vercel/analytics';
 
@@ -8,11 +8,6 @@ import BrowserWindow from './components/browser-window/browser-window.component'
 import IconBar from './components/icon-bar/icon-bar.component';
 import Explorer from './components/explorer/explorer.component';
 import StatusBar from './components/status-bar/status-bar.component';
-
-import Home from './routes/home/home.component';
-import Projects from './routes/projects/projects.component';
-import Resume from './routes/resume/resume.component';
-import Stats from './routes/stats/stats.component';
 
 function App() {
   // VERCEL ANALYTICS
@@ -24,12 +19,7 @@ function App() {
       <div className="app">
         <IconBar />
         <Explorer />
-        <Routes>
-          <Route path='/' element={<Home />} />
-          <Route path='projects/*' element={<Projects />} />
-          <Route path='resume' element={<Resume />} />
-          <Route path='stats' element={<Stats />} />
-        </Routes>
+        <Outlet />
       </div>
       <StatusBar />
     </div>
