+++ src/components/project-page/project-page.styles.scss
@@ -2,20 +2,16 @@
     background-color: #1D1D2E;
     width: 100%;
     z-index:100;
-    padding: 1% 2%;
-    font-size: 0.9em;
     overflow-y: auto;
     overflow-x: hidden;
     height: 100%;
-    font-size: 1em;
-    padding-bottom: 50px;
 
     .md{
         .markdown-body{
             
         }
         h1:first-child{
-            margin-top: 1%;
+            margin-top: 0;
         }
         h1, h2, h3{
             margin: 3% 0 1%;
